+++ src/core/byte-cursor.ts
@@ -55,6 +55,7 @@ export class ByteCursor {
  * @throws {BareError} bytes are missing.
  */
 export function check(bc: ByteCursor, min: number): void {
+    // We try to keep this function as small as possible to allow inling.
     if (DEV) {
         assert(isU32(min))
     }
@@ -69,29 +70,91 @@ export function check(bc: ByteCursor, min: number): void {
  * @throws {BareError} Buffer exceeds `config.maxBufferLength`.
  */
 export function reserve(bc: ByteCursor, min: number): void {
+    // We try to keep this function as small as possible to allow inling.
     if (DEV) {
         assert(isU32(min))
     }
     const minLen = (bc.offset + min) | 0
     if (minLen > bc.bytes.length) {
-        if (minLen > bc.config.maxBufferLength) {
-            throw new BareError(0, TOO_LARGE_BUFFER)
-        }
-        //
-        // |     bytes,view}.buffer         |
-        //     |        bytes        |
-        //     |        view         |
-        //     |<-- offset -->|<-- min -->|
-        //     |<-------- minLen -------->|
-        //     |         new view                         |
-        //     |         new bytes                        |
-        //     |         new {bytes,view}.buffer          |
-        //     |<------------- config.maxBufferLength -------------->|
-        //
-        const newLen = Math.min(minLen << 1, bc.config.maxBufferLength)
-        const newBytes = new Uint8Array(newLen)
-        newBytes.set(bc.bytes)
-        bc.bytes = newBytes
-        bc.view = new DataView(newBytes.buffer)
+        grow(bc, minLen)
     }
+}
+
+/**
+ * Grow the underlying buffer of `bc` such that its length is
+ * greater or equal to `minLen`.
+ *
+ * @throws {BareError} Buffer exceeds `config.maxBufferLength`.
+ */
+function grow(bc: ByteCursor, minLen: number): void {
+    if (minLen > bc.config.maxBufferLength) {
+        throw new BareError(0, TOO_LARGE_BUFFER)
+    }
+    //
+    // |     bytes,view}.buffer         |
+    //     |        bytes        |
+    //     |        view         |
+    //     |<-- offset -->|<-- min -->|
+    //     |<-------- minLen -------->|
+    //     |         new view                         |
+    //     |         new bytes                        |
+    //     |         new {bytes,view}.buffer          |
+    //     |<------------- config.maxBufferLength -------------->|
+    //
+    const buffer = bc.bytes.buffer
+    let newBytes: Uint8Array
+    if (
+        isEs2024ArrayBufferLike(buffer) &&
+        // Make sure that the view covers the end of the buffer.
+        // If it is not the case, this indicates that the user don't want
+        // to override the trailing bytes.
+        bc.bytes.byteOffset + bc.bytes.byteLength === buffer.byteLength &&
+        bc.bytes.byteLength + minLen <= buffer.maxByteLength
+    ) {
+        const newLen = Math.min(
+            minLen << 1,
+            bc.config.maxBufferLength,
+            buffer.maxByteLength,
+        )
+        if (buffer instanceof ArrayBuffer) {
+            buffer.resize(newLen)
+        } else {
+            buffer.grow(newLen)
+        }
+        newBytes = new Uint8Array(buffer, bc.bytes.byteOffset, newLen)
+    } else {
+        const newLen = Math.min(minLen << 1, bc.config.maxBufferLength)
+        newBytes = new Uint8Array(newLen)
+        newBytes.set(bc.bytes)
+    }
+    bc.bytes = newBytes
+    bc.view = new DataView(newBytes.buffer)
+}
+
+interface Es2024ArrayBuffer extends ArrayBuffer {
+    readonly detached: boolean
+    readonly resizable: boolean
+    readonly maxByteLength: number
+    /**
+     * @throws {TypeError} if the buffer is `detached` or is not `resizable`.
+     * @throws {RangeError} if `newLength` is larger than `maxByteLength`.
+     */
+    resize(newLength: number): void
+}
+
+interface Es2024SharedArrayBuffer extends SharedArrayBuffer {
+    readonly growable: boolean
+    readonly maxByteLength: number
+    /**
+     * @throws {TypeError} if the buffer is not `growable`.
+     * @throws {RangeError} if `newLength` is larger than `maxByteLength`
+     *                      or smaller than `byteLength`.
+     */
+    grow(newLength: number): void
+}
+
+function isEs2024ArrayBufferLike(
+    buffer: ArrayBufferLike,
+): buffer is Es2024ArrayBuffer | Es2024SharedArrayBuffer {
+    return "maxByteLength" in buffer
 }